Introduce src/utils/ for utility function and make them more generic
