+++ src/widget/index.tsx
@@ -1,6 +1,8 @@
 import {h, render} from 'preact';
 import Widget from './widget';
 import {defaultConfiguration} from './configuration';
+import { getUrlParameter } from '../utils';
+import { IConfiguration } from '../typings';
 
 if (window.attachEvent) {
     window.attachEvent('onload', injectChat);
@@ -8,12 +10,7 @@ if (window.attachEvent) {
     window.addEventListener('load', injectChat, false);
 }
 
-function getUrlParameter(name: string, defaults = '') {
-    name = name.replace(/[[]/, '\\[').replace(/[]]/, '\\]');
-    let regex = new RegExp('[\\?&]' + name + '=([^&#]*)');
-    let results = regex.exec(document.getElementById('botmanWidget').getAttribute('src'));
-    return results === null ? defaults : decodeURIComponent(results[1].replace(/\+/g, ' '));
-}
+
 function injectChat() {
     let root = document.createElement('div');
     root.id = 'botmanWidgetRoot';
@@ -21,7 +18,7 @@ function injectChat() {
 
     let settings = {};
     try {
-        settings = JSON.parse(getUrlParameter('settings', '{}'));
+        settings = JSON.parse(getUrlParameter(document.getElementById('botmanWidget').getAttribute('src'), 'settings', '{}'));
     } catch (e) { }
 
     const dynamicConf = window.botmanWidget || {}; // these configuration are loaded when the chat frame is opened
@@ -42,5 +39,5 @@ function injectChat() {
 }
 
 declare global {
-    interface Window { attachEvent: Function, botmanWidget: Widget }
+    interface Window { attachEvent: Function, botmanWidget: IConfiguration }
 }
